fascinatemtd: remove libakm.so from proprietary files list
libakm.so is not present on any of the CM images and does not appear
to be needed.  Removing it from the list makes it possible to build a
new fascinatemtd image from an existing CyanogenMod image.

If needed, the Replicant project apparently has a GPL3-licenced
reimplementation of libakm:
https://gitorious.org/replicant/hardware_libakm_free

Change-Id: I5284f67db99b28e080660e7d9b41021127660326